Pass device=cpu to torch.asarray()
The code in get_tensor assumes that torch.asarray will always load the
tensor in cpu, and then the tensor is copied to self.device in the "to"
call at the end of the function. However, that assumption is not always
correct. To make sure, device=cpu has to be passed to asarray.