Fix release CI issues (#673)
* Fix release CI issues

- Switch to manylinux_2_28.
- Drop Python 3.8, add 3.12 and 3.13.
- Fix miniconda failures.

The motivation to switch to manylinux_2_28 is because manylinux2014 is
starting to break. E.g. the container cannot run GitHub actions like
`checkout` anymore because the NodeJS version was updated and Miniconda
cannot be installed in the container because it does not support the old
glibc version anymore.

manylinux_2_28 uses glibc 2.28. This version is supported by Red Hat
Enterprise Linux 8 (2019) and Ubuntu 20.04.

Also see:

https://github.com/actions/checkout/issues/1809
https://www.anaconda.com/docs/getting-started/miniconda/system-requirements

* CI: macos-13 runner is deprecated